Revert "COST-3067 OCI Ingestion for day after today (#414)"
This reverts commit b4a185b1578b78713c5726ff25e7f5949e06122a.

+++ tests/test_oci_generator.py
@@ -178,8 +178,8 @@ class TestOCIComputeGenerator(OCIGeneratorTestCase):
         """Test that usage data is added."""
         generator = OCIComputeGenerator(self.six_hours_ago, self.now, self.currency, self.attributes)
         usage_row = generator._add_usage_data({}, self.six_hours_ago, self.now)
-        self.assertIsInstance(usage_row["usage/consumedQuantity"], int)
-        self.assertIsInstance(usage_row["usage/billedQuantity"], int)
+        self.assertIsInstance(usage_row["usage/consumedQuantity"], float)
+        self.assertIsInstance(usage_row["usage/billedQuantity"], float)
         self.assertIsInstance(usage_row["usage/consumedQuantityUnits"], str)
         self.assertIsInstance(usage_row["usage/consumedQuantityMeasure"], str)
 
@@ -218,8 +218,8 @@ class TestOCINetworkGenerator(OCIGeneratorTestCase):
         """Test that usage specific data for network service is added."""
         generator = OCINetworkGenerator(self.six_hours_ago, self.now, self.currency, self.attributes)
         usage_row = generator._add_usage_data({}, self.six_hours_ago, self.now)
-        self.assertIsInstance(usage_row["usage/consumedQuantity"], int)
-        self.assertIsInstance(usage_row["usage/billedQuantity"], int)
+        self.assertIsInstance(usage_row["usage/consumedQuantity"], float)
+        self.assertIsInstance(usage_row["usage/billedQuantity"], float)
         self.assertIsInstance(usage_row["usage/consumedQuantityUnits"], str)
         self.assertIsInstance(usage_row["usage/consumedQuantityMeasure"], str)
 
@@ -252,8 +252,8 @@ class TestOCIBlockStorageGenerator(OCIGeneratorTestCase):
         """Test that usage specific data for block storage service is added."""
         generator = OCIBlockStorageGenerator(self.six_hours_ago, self.now, self.currency, self.attributes)
         test_row = generator._add_usage_data({}, self.six_hours_ago, self.now)
-        self.assertIsInstance(test_row["usage/consumedQuantity"], int)
-        self.assertIsInstance(test_row["usage/billedQuantity"], int)
+        self.assertIsInstance(test_row["usage/consumedQuantity"], float)
+        self.assertIsInstance(test_row["usage/billedQuantity"], float)
         self.assertIsInstance(test_row["usage/consumedQuantityUnits"], str)
         self.assertIsInstance(test_row["usage/consumedQuantityMeasure"], str)
 
@@ -285,7 +285,7 @@ class TestOCIDatabaseGenerator(OCIGeneratorTestCase):
         """Test that usage specific data for database service is added."""
         generator = OCIDatabaseGenerator(self.two_hours_ago, self.now, self.currency, self.attributes)
         test_row = generator._add_usage_data({}, self.two_hours_ago, self.now)
-        self.assertIsInstance(test_row["usage/consumedQuantity"], int)
-        self.assertIsInstance(test_row["usage/billedQuantity"], int)
+        self.assertIsInstance(test_row["usage/consumedQuantity"], float)
+        self.assertIsInstance(test_row["usage/billedQuantity"], float)
         self.assertIsInstance(test_row["usage/consumedQuantityUnits"], str)
         self.assertIsInstance(test_row["usage/consumedQuantityMeasure"], str)
